+++ pangea-sdk/v3/service/audit/roots_providers.go
@@ -11,6 +11,7 @@ import (
 
 type RootsProvider interface {
 	UpdateRoots(ctx context.Context, treeSizes []string) map[int]Root
+	OverrideRoots(roots map[int]Root) map[int]Root
 }
 
 type ArweaveRootsProvider struct {
@@ -67,6 +68,13 @@ func (rp *ArweaveRootsProvider) UpdateRoots(ctx context.Context, treeSizes []str
 			}
 			rp.Roots[size] = root
 		}
+	}
+	return rp.Roots
+}
+
+func (rp *ArweaveRootsProvider) OverrideRoots(roots map[int]Root) map[int]Root {
+	for treeSize, root := range roots {
+		rp.Roots[treeSize] = root
 	}
 	return rp.Roots
 }