+++ pangea-sdk/v3/service/vault/common.go
@@ -19,6 +19,7 @@ const (
 	KPsigning    KeyPurpose = "signing"
 	KPencryption KeyPurpose = "encryption"
 	KPjwt        KeyPurpose = "jwt"
+	KPfpe        KeyPurpose = "fpe" // Format-preserving encryption.
 )
 
 type AsymmetricAlgorithm string
@@ -65,15 +66,17 @@ const (
 type SymmetricAlgorithm string
 
 const (
-	SYAhs256      SymmetricAlgorithm = "HS256"
-	SYAhs384      SymmetricAlgorithm = "HS384"
-	SYAhs512      SymmetricAlgorithm = "HS512"
-	SYAaes128_cfb SymmetricAlgorithm = "AES-CFB-128"
-	SYAaes256_cfb SymmetricAlgorithm = "AES-CFB-256"
-	SYAaes256_gcm SymmetricAlgorithm = "AES-GCM-256"
-	SYAaes128_cbc SymmetricAlgorithm = "AES-CBC-128"
-	SYAaes256_cbc SymmetricAlgorithm = "AES-CBC-256"
-	SYAaes        SymmetricAlgorithm = "AES-CFB-128" // deprecated, use SYAaes128_cfb instead
+	SYAhs256         SymmetricAlgorithm = "HS256"
+	SYAhs384         SymmetricAlgorithm = "HS384"
+	SYAhs512         SymmetricAlgorithm = "HS512"
+	SYAaes128_cfb    SymmetricAlgorithm = "AES-CFB-128"
+	SYAaes256_cfb    SymmetricAlgorithm = "AES-CFB-256"
+	SYAaes256_gcm    SymmetricAlgorithm = "AES-GCM-256"
+	SYAaes128_cbc    SymmetricAlgorithm = "AES-CBC-128"
+	SYAaes256_cbc    SymmetricAlgorithm = "AES-CBC-256"
+	SYAaes           SymmetricAlgorithm = "AES-CFB-128"        // deprecated, use SYAaes128_cfb instead
+	SYAaes_ff3_1_128 SymmetricAlgorithm = "AES-FF3-1-128-BETA" // 128-bit encryption using the FF3-1 algorithm. Beta feature.
+	SYAaes_ff3_1_256 SymmetricAlgorithm = "AES-FF3-1-256-BETA" // 256-bit encryption using the FF3-1 algorithm. Beta feature.
 )
 
 type ItemVersionState string
@@ -129,6 +132,12 @@ const (
 	IOBfolder       ItemOrderBy = "folder"
 	IOBversion      ItemOrderBy = "version"
 )
+
+// Algorithm of an exported public key.
+type ExportEncryptionAlgorithm string
+
+// RSA 4096-bit key, OAEP padding, SHA512 digest.
+const EEArsa4096_oaep_sha512 ExportEncryptionAlgorithm = "RSA-OAEP-4096-SHA512"
 
 type CommonStoreRequest struct {
 	// Base request has ConfigID for multi-config projects
@@ -206,6 +215,7 @@ type ItemData struct {
 	CreatedAt         string          `json:"created_at"`
 	Algorithm         string          `json:"algorithm,omitempty"`
 	Purpose           string          `json:"purpose,omitempty"`
+	Exportable        *bool           `json:"exportable,omitempty"` // Whether the key is exportable or not.
 }
 
 type InheritedSettings struct {
@@ -333,4 +343,25 @@ type FolderCreateRequest struct {
 
 type FolderCreateResult struct {
 	ID string `json:"id"`
+}
+
+type ExportRequest struct {
+	pangea.BaseRequest
+
+	ID                  string                     `json:"id"`                             // The ID of the item.
+	Version             *int                       `json:"version,omitempty"`              // The item version.
+	EncryptionKey       *string                    `json:"encryption_key,omitempty"`       // Public key in PEM format used to encrypt exported key(s).
+	EncryptionAlgorithm *ExportEncryptionAlgorithm `json:"encryption_algorithm,omitempty"` // The algorithm of the public key.
+}
+
+type ExportResult struct {
+	ID         string  `json:"id"`                    // The ID of the item.
+	Version    int     `json:"version"`               // The item version.
+	Type       string  `json:"type"`                  // The type of the key.
+	ItemState  string  `json:"item_state"`            // The state of the item.
+	Algorithm  string  `json:"algorithm"`             // The algorithm of the key.
+	PublicKey  *string `json:"public_key,omitempty"`  // The public key (in PEM format).
+	PrivateKey *string `json:"private_key,omitempty"` // The private key (in PEM format).
+	Key        *string `json:"key,omitempty"`         // The key material.
+	Encrypted  bool    `json:"encrypted"`             // Whether exported key(s) are encrypted with encryption_key sent on the request or not. If encrypted, the result is sent in base64, any other case they are in PEM format plain text.
 }